add skip mark
