+++ tests/test_benchmarks/test_benchmark_score.py
@@ -1,7 +1,4 @@
 import numpy as np
-import polars as pl
-import pytest
-from datasets import Dataset
 
 import mteb
 from mteb import ResultCache
@@ -13,30 +10,7 @@ from mteb.benchmarks._benchmark_metrics import (
     _compute_task_types,
 )
 from mteb.benchmarks.benchmark import Benchmark, BenchmarkAggregation
-
-
-def _datasets_supports_dictionary_type() -> bool:
-    """True if the installed ``datasets`` can convert Polars Categorical columns.
-
-    ``_to_results_df`` goes through ``Dataset.from_polars`` for categorical
-    columns (model_name, task_name, …); older ``datasets`` releases raise
-    ``NotImplementedError`` for ``pa.DictionaryType``. Probe once at import.
-    """
-    try:
-        Dataset.from_polars(pl.DataFrame({"x": ["a"]}, schema={"x": pl.Categorical}))
-    except NotImplementedError:
-        return False
-    return True
-
-
-_skip_if_datasets_too_old = pytest.mark.skipif(
-    not _datasets_supports_dictionary_type(),
-    reason=(
-        "installed `datasets` cannot convert Polars Categorical columns "
-        "(pa.DictionaryType -> Features.from_arrow_schema NotImplementedError); "
-        "skip the _to_results_df-based parity tests on lowest-pin CI"
-    ),
-)
+from tests.conftest import _skip_if_datasets_too_old
 
 MODELS_SCORES = {
     "Mean(Task)": {
